renamed dice to dice_logic, renamed maexchen to maexchen_logic
